pyproject.toml
[tool.poetry]
name = "12 - desenvolvimento fullstack com django"
version = "0.1.0"
description = ""
authors = ["Nicholas <sousa.nicholas2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "*"
dj-database-url = "*"
psycopg2-binary = "*"
whitenoise = {version = "*", extras = ["brotli"]}
gunicorn = "*"
waitress = "^3.0.2"


[tool.poetry.group.dev.dependencies]
ipython = "*"
pytest-django = "*"


[tool.flake8]
max-line-length = 300


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"